deviding getNews & getPickUpNews#1
